requirements.txt
streamlit>=1.44
pandas>=2.2
requests>=2.32
feedparser>=6.0
python-dateutil>=2.9
PyYAML>=6.0
beautifulsoup4>=4.12
lxml>=5.2
